Front/ast.cpp: copy formal arguments when creating Code_Import
This fixes a bug which emerged when an import was used in a predicate
called from more than one place.  After the first use, the formal
arguments of the import were deleted and the second access was reading
invalid memory, which resulted in segmentation fault.

The fix copies the formal arguments for every instance of the import.
Note that a small memory leak (of the original formal arguments) may
have been introduced.